redis: Split into multiple files
