+++ src/redis/executor.rs
@@ -1,0 +1,17 @@
+use actix::prelude::*;
+use r2d2::Pool;
+use r2d2_redis::RedisConnectionManager;
+
+pub struct RedisExecutor {
+    pub pool: Pool<RedisConnectionManager>,
+}
+
+impl RedisExecutor {
+    pub fn new(pool: Pool<RedisConnectionManager>) -> Self {
+        RedisExecutor { pool }
+    }
+}
+
+impl Actor for RedisExecutor {
+    type Context = SyncContext<Self>;
+}
